ai-support-agent/functions/requirements.txt
firebase-functions
firebase-admin
google-api-python-client
google-auth
mailersend
google-generativeai
python-dotenv